refact : actions & command
