+++ src/bin/dummy.ts
@@ -1,11 +1,7 @@
 #!/usr/bin/env node
-import { GenerateCommand } from "../commands";
-import { BANNER } from "../ui";
-const init = new GenerateCommand
+import { GenerateCommand } from '../commands';
+import { callDummyGenearteBanner } from '../ui';
+const init = new GenerateCommand();
 
-
-console.log(BANNER)
-init.load()
-
-
-
+callDummyGenearteBanner();
+init.loadCommands();
