+++ src/commands/actions/index.ts
@@ -1,0 +1,3 @@
+// export { InitAction } from './init.action';
+export { initCommand } from './init.action';
+export { generateCommand } from './generate.action';
