+++ src/ui/banner.ui.ts
@@ -1,3 +1,4 @@
-const figlet = require('figlet'); 
+const figlet = require('figlet');
 
-export const BANNER = figlet.textSync("Dummy Generator");
+export const callDummyGenearteBanner = () =>
+  console.log(figlet.textSync('Dummy Generator'));
